+++ config.inc.php
@@ -16,11 +16,11 @@ $REX['EXTPERM'][] = 'seo42[seo_extended]';
 $REX['EXTPERM'][] = 'seo42[url_default]';
 
 // includes
-require($REX['INCLUDE_PATH'] . '/addons/seo42/classes/class.seo42.inc.php');
-require($REX['INCLUDE_PATH'] . '/addons/seo42/classes/class.seo42_utils.inc.php');
-require($REX['INCLUDE_PATH'] . '/addons/seo42/settings.dyn.inc.php');
-require($REX['INCLUDE_PATH'] . '/addons/seo42/settings.advanced.inc.php');
-require($REX['INCLUDE_PATH'] . '/addons/seo42/settings.lang.inc.php');
+require_once($REX['INCLUDE_PATH'] . '/addons/seo42/classes/class.seo42.inc.php');
+require_once($REX['INCLUDE_PATH'] . '/addons/seo42/classes/class.seo42_utils.inc.php');
+require_once($REX['INCLUDE_PATH'] . '/addons/seo42/settings.dyn.inc.php');
+require_once($REX['INCLUDE_PATH'] . '/addons/seo42/settings.advanced.inc.php');
+require_once($REX['INCLUDE_PATH'] . '/addons/seo42/settings.lang.inc.php');
 
 // fix for iis webserver: set request uri manually if not available
 if (!isset($_SERVER['REQUEST_URI'])) {